refactor(tests): separate tests into multiple files
