+++ tests/index.test.ts
@@ -1,0 +1,22 @@
+/*
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+
+import mdl, { MagentoDataLayer } from "../src/index";
+
+beforeAll(() => {
+  // Forces magento data layer code to be bundled so that
+  // 'data layer should exist' test passes
+  expect(mdl).toBeInstanceOf(MagentoDataLayer);
+});
+
+beforeEach(async () => {
+  jest.resetModules();
+  window.adobeDataLayer = [];
+  require("@adobe/adobe-client-data-layer");
+});
+
+test("data layer should exist", () => {
+  expect(window.adobeDataLayer).toBeDefined();
+});
